tools/theme/requirements.txt
pefile
pillow
